fixed table, made it to be scrollable at a fixed height and width, made it dynamic
