+++ package.json
@@ -14,6 +14,7 @@
     "@nextui-org/react": "^2.2.10",
     "@nextui-org/system": "^2.0.15",
     "@nextui-org/theme": "^2.1.18",
+    "@nextui-org/use-infinite-scroll": "^2.1.3",
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-slot": "^1.0.2",
     "@react-stately/data": "^3.11.2",
